made auth function for server side
