+++ app/(pages)/newsLetter/page.tsx
@@ -1,3 +1,3 @@
-export default function NewsLetter() {
+export default async function NewsLetter() {
   return <div>NewsLetter</div>;
 }
