+++ app/api/auth/user/route.ts
@@ -4,10 +4,10 @@ import { db } from "@/db";
 import { createSession, decrypt } from "@/lib/session";
 import { SessionPayload, SessionUser } from "@/types";
 import { cookies } from "next/headers";
-import { NextResponse } from "next/server";
+import {  NextRequest, NextResponse } from "next/server";
 
-export async function GET() {
-  const token = (await cookies()).get("session")?.value;
+export async function GET(request: NextRequest) {
+  const token = (await cookies()).get("session")?.value || request.headers.get("Authorization")?.split("Bearer ")[1];
   if (!token) {
     return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
   }
